Added a method to scrape dynamic(javascript) website
The method works. But it will open up chrome on the machine. It's unimaginable if thousands of requests are made to scrape the website. So we still need to find out other solutions for scraping dynamic website.

And I changed to reading communites by the order of price. It's much more efficient than default order.